requirements.txt
imageio==2.33.1
joblib==1.3.2
numpy==1.26.3
opencv-python==4.9.0.80
pandas==2.1.4
pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.4
py5~=0.10.0a0